BespinApp/requirements.txt
django
djangorestframework
markdown
django-filter
django-grappelli==2.6.4
django-celery
twilio
